+++ gulpfile.js
@@ -9,8 +9,13 @@ const cleanCSS = require('gulp-clean-css');
 const rename = require('gulp-rename');
 const babel = require('gulp-babel');
 const mergeStream = require('merge-stream');
+const ENV = process.env.NODE_MODULE_ENV;
 
 /**
+ * TODO: When the packaging fails, the scene needs to be restored
+ */
+/**
+
  * existStyleCatalogName ---- The name of the component catalog where the style exists
  * componentCatalogName ---- All component catalog names
  * folderName ---- JavaScript temporary folder name
@@ -21,21 +26,27 @@ const folderName = '.temporary_files';
 const basePath = 'src/components/';
 
 function difference(array, arrCompare) {
-  return array.concat(arrCompare).filter(function(v, i, arr) {
+  return array.concat(arrCompare).filter(function (v, i, arr) {
     return arr.indexOf(v) === arr.lastIndexOf(v);
   });
 }
+
 function unique(arr) {
   return Array.from(new Set(arr));
 }
 
 function writeStyleFile(catalogArr) {
   for (let catalogName of catalogArr) {
-    fs.writeFile(`${basePath}${catalogName}/style.scss`, '', { flag: 'wx' }, err => {
-      if (err) {
-        throw err;
-      }
-    });
+    fs.writeFile(
+      `${basePath}${catalogName}/style.scss`,
+      '',
+      { flag: 'wx' },
+      (err) => {
+        if (err) {
+          throw err;
+        }
+      },
+    );
   }
 }
 
@@ -44,6 +55,7 @@ function delStyleFile(catalogArr) {
     del([`${basePath}${catalogName}/style.scss`]);
   }
 }
+
 /**
  * traverse files
  * @param isDelete Distinguish between write and delete style file
@@ -51,7 +63,7 @@ function delStyleFile(catalogArr) {
 function traverseExistStyleFile(isDelete) {
   return src([`${basePath}*/*.scss`])
     .pipe(
-      through.obj(function(file, enc, callback) {
+      through.obj(function (file, enc, callback) {
         isExist = Boolean(file.contents.toString());
         if (isExist) {
           existStyleCatalogName.push(file.relative.split('/')[0]);
@@ -59,8 +71,11 @@ function traverseExistStyleFile(isDelete) {
         callback();
       }),
     )
-    .on('end', async function() {
-      const noStyleComp = difference(unique(componentCatalogName), unique(existStyleCatalogName));
+    .on('end', async function () {
+      const noStyleComp = difference(
+        unique(componentCatalogName),
+        unique(existStyleCatalogName),
+      );
       if (isDelete) {
         delStyleFile(noStyleComp);
         console.log('delStyleFile exec end');
@@ -76,12 +91,12 @@ function traverseComponent() {
   console.warn('Do not edit the source file when the project is compiled !!!');
   return src([`${basePath}*/`, `!${basePath}utils/`]) // exclude utils/
     .pipe(
-      through.obj(function(file, enc, callback) {
+      through.obj(function (file, enc, callback) {
         componentCatalogName.push(file.relative.split('/')[0]);
         callback();
       }),
     )
-    .on('end', function() {
+    .on('end', function () {
       console.log('traverseComponent exec end ~');
     });
 }
@@ -99,17 +114,32 @@ function cleanUselessStyleFile() {
 }
 
 async function descriptionDoc() {
-  await fs.mkdir(folderName, { recursive: false }, err => {
+  await fs.mkdir(folderName, { recursive: false }, (err) => {
     if (err) throw err;
   });
-  await fs.writeFile(`${folderName}/index.tsx`, `import './style.scss';`, { flag: 'w' }, err => {
-    if (err) {
-      throw err;
-    }
-    console.log(`Do not delete the ${folderName} until the compilation is complete`);
-  });
+  await fs.writeFile(
+    `${folderName}/index.tsx`,
+    `import './style.scss';`,
+    { flag: 'w' },
+    (err) => {
+      if (err) {
+        throw err;
+      }
+      console.log(
+        `Do not delete the ${folderName} until the compilation is complete`,
+      );
+    },
+  );
   return src(`${folderName}/index.tsx`)
-    .pipe(ts({ declaration: true, target: 'ES5' }))
+    .pipe(
+      ts({
+        declaration: true,
+        target: 'ES5',
+        module: ENV === 'esm' ? 'es6' : 'commonjs',
+        skipLibCheck: true,
+        allowSyntheticDefaultImports: true
+      }),
+    )
     .pipe(dest(`${folderName}/`));
 }
 
@@ -118,7 +148,7 @@ function outputStyleTask(componentCatalogName) {
     Object.prototype.toString.call(componentCatalogName) !== '[Object Array]' &&
     componentCatalogName.length !== 0
   ) {
-    componentCatalogName.map(componentName => {
+    componentCatalogName.map((componentName) => {
       convertStyles(componentName);
     });
     return Promise.resolve('Components style file output completed');
@@ -129,51 +159,52 @@ function outputStyleTask(componentCatalogName) {
 function globalSass() {
   return src(`${basePath}**/*.scss`)
     .pipe(concat('index.scss'))
-    .pipe(dest('lib/style'));
+    .pipe(dest(ENV + '/style'));
 }
 
 async function clean(cb) {
   await cleanUselessStyleFile();
-  await del(['lib', '.temporary_files']);
+  await del([ENV, '.temporary_files']);
   await cb();
 }
 
 function globalCss() {
   del(folderName);
-  return src('lib/**/*.scss')
+  return src(ENV + '/**/*.scss')
     .pipe(concat('index.css'))
     .pipe(sass())
     .on('error', sass.logError)
     .pipe(cleanCSS({ compatibility: 'ie11' }))
-    .pipe(dest('lib/style'));
+    .pipe(dest(ENV + '/style'));
 }
 
 function convertStyles(data) {
   return mergeStream(
     outputForCss(data),
-    outputForSaSS(data,'index.js'),
-    outputForSaSS(data,'index.d.ts'),
-    outputForCssFile(data)
+    outputForSaSS(data, 'index.js'),
+    outputForSaSS(data, 'index.d.ts'),
+    outputForCssFile(data),
   );
 }
 
-function outputForSaSS(data, fileName){
-  return src(`${folderName}/${fileName}`)
-    .pipe(dest('lib/' + String(data) + '/style/'))
+function outputForSaSS(data, fileName) {
+  return src(`${folderName}/${fileName}`).pipe(
+    dest(ENV + '/' + String(data) + '/style/'),
+  );
 }
 
-function outputForCssFile(data){
+function outputForCssFile(data) {
   return src(`${folderName}/index.tsx`)
-  .pipe(rename('css.tsx'))
-  .pipe(babel())
-  .pipe(dest('lib/' + String(data) + '/style/'))
+    .pipe(rename('css.tsx'))
+    .pipe(babel())
+    .pipe(dest(ENV + '/' + String(data) + '/style/'));
 }
 
 function outputForCss(data) {
   return src([`${basePath}` + String(data) + '/*.scss'])
-  .pipe(dest('lib/' + String(data) + '/style/'))
-  .pipe(sass())
-  .pipe(dest('lib/' + String(data) + '/style/'))
+    .pipe(dest(ENV + '/' + String(data) + '/style/'))
+    .pipe(sass())
+    .pipe(dest(ENV + '/' + String(data) + '/style/'));
 }
 
 exports.default = series(