+++ src/components/index.tsx
@@ -1,15 +1,16 @@
-
-export { default as BreadcrumbRender } from './breadcrumb';
+export { default as Breadcrumb } from './breadcrumb';
 export { default as ChromeDownload } from './chromeDownload';
 export { default as Circle } from './circle';
 export { default as ContextMenu } from './contextMenu';
 export { default as Cookies } from './cookies';
+export { default as ContextMenuCombiner } from './ContextMenuCombiner';
+export { default as CopyIcon } from './copyIcon';
 export { default as EasySelect } from './easySelect';
 export { default as EditCell } from './editCell';
 export { default as EditInput } from './editInput';
 export { default as EllipsisText } from './ellipsisText';
 export { default as ErrorBoundary } from './errorBoundary';
-export { default as FullScreenButton } from './fullscreen';
+export { default as Fullscreen } from './fullscreen';
 export { default as GlobalLoading } from './globalLoading';
 export { default as GoBack } from './goBack';
 export { default as KeyEventListener } from './keyEventListener';
@@ -20,7 +21,7 @@ export { default as MulSelectDropdown } from './mulSelectDropdown';
 export { default as MultiSearchInput } from './multiSearchInput';
 export { default as NotFound } from './notFound';
 export { default as ProgressBar } from './progressBar';
-export { default as RenderFormItem } from './formComponent';
+export { default as RenderFormItem } from './renderFormItem';
 export { default as Resize } from './resize';
 export { default as ScrollText } from './scrollText';
 export { default as SearchModal } from './searchModal';
@@ -30,4 +31,5 @@ export { default as SwitchWindow } from './switchWindow';
 export { default as TableCell } from './tableCell';
 export { default as TextMark } from './textMark';
 export { default as ToolModal } from './toolModal';
+export { default as ProgressLine } from './progressLine'